Create authorization functionality
Added authorization functionality that enables the user to sign in,
sign up, sign out, and change password. All of the functions are
contained in a navbar at the top and only appear when their function
is warranted. Also changed the api.origin url paths in
assets/scripts/config.js and the project name in our package.json file.