Added the index for pages
First added the button for the getting the index of pages. Then we went
to the the api to set up the actually index function. After we went
to the events and created a on Index function to allow the flow of
response of either success or error and what to display depending on the
response. We defined what those responses should be in the UI and then
set up the click handler to receive input from the user. Then tested to
make sure we got a response from the server, and that it was the correct
response.
Next step we're moving on to the show and destroy functions.