Finished the show for Pages
Made the show for pages on the site. Started with the api and the
events. Constructed the showpages in the api, then constructed the
onshowpages in the events. These allow the input of the id to find a
specific page in the list of created pages. Using the success function
from the UI it states when there is no ID passed.

Next we're going to make the destroy function for pages.