Added the crud actions for Blog
First we changed over all the crud commands and scripts for the blog
Then we added some click handlers in the event.js for crud functions

Next we're going fix an addition in the html to allow complete
crud for the blog, we're only able to crud the title. We're going to add
the content crudability.