Implement  auto sign-in with sign-up
Added the sign-in to sign-up in the auth/events.js so that when
the sign-up event fires it automatically logs the user in.
The function works, however in the chrome console there is a warning
that says:

'jQuery.Deferred exception: Cannot read property 'then' of undefined
TypeError: Cannot read property 'then' of undefined
at Object.<anonymous>'

Not sure what this means, but will be filing an issue about this next.

This should not affect anything else and will be making pull request
to  merge into dev2.