Implementation of handle bars for pages
In index.js click event listeners have been modified for delete
update for preventing bubbling issues.

On api.js destroyPage and Update page has been modified so that
we can get id and data out of events and send it to database

ui.js in update required handle bars from their origin to render
the data when the action is on success. Also in remove, when delete
is successful specific div that exist on html is delete.

page-crud.handlebars are modified into models so that they can match
with the wireframes and it can imply functionalty

Page-listing fade class of modals are deleted so that it could not create
any issues when the action happens

update-page is created to over-write the pre existing data