Add blog listing function and associated crud actions
Created the ability to list each individual blog and perform crud action
on each that were listed, granted the user logged in is the owner of
said blog. Also added associated listeners so that when a blog
is updated or deleted the whole list refreshes to update changes.
Also initiated the ability to use the TinyMCE plugin.