Refactor Homepage
Refactored homepage so that when a user is not signed in and they click
the index blogs button it shows all of the blogs for all users without
edit and delete buttons. Once a user logs in and they click the same
button, the blogs are listed to include edit and delete buttons.
Next I will be tidying up the code to make elements hide and show when
proper.