Update user interface, esp. regarding page resource
Updated the user interface so when a user first visits the page he/she
cannot view or edit pages. Only when a user signs in can he/she do so.
Also make it so the create and show webpage buttons disappear when the
user signs out. Next we will need to finish formatting the input fields
for both blog and pages, including autopopulating input fields, and
clearing input fields when a successful resource has been created.